+++ src/schema.gql
@@ -49,8 +49,8 @@ type MoviesResponseGQL {
 }
 
 type Query {
-  getAllGenresByUserId(userId: Int!): [FavoriteGenre!]!
-  getAllMoviesByUserId(userId: Int!): [FavoriteMovie!]!
+  getUserMovies(userId: Int!): [FavoriteMovie!]!
+  getUserGenres(userId: Int!): [FavoriteGenre!]!
   getAllGenres(language: String): [GenreGQL!]!
   getMoviesByFilters(language: String, genreIds: [Int!], minCountVotes: Int, releaseYear: Int, page: Int): MoviesResponseGQL!
   getFavoriteMoviesByIds(ids: [Int!]!, language: String, page: Int): MoviesResponseGQL!
@@ -59,10 +59,9 @@ type Query {
 }
 
 type Mutation {
-  addGenreIdToUserByUserId(userId: Int!, genreId: Int!): FavoriteGenre!
-  removeGenreIdToUserByUserId(userId: Int!, genreId: Int!): String!
-  addMovieIdToUserByUserId(userId: Int!, movieId: Int!): FavoriteMovie!
-  removeMovieIdToUserByUserId(userId: Int!, movieId: Int!): String!
-  watchByMovieIdToUserByUserId(userId: Int!, movieId: Int!): FavoriteMovie!
-  unWatchByMovieIdToUserByUserId(userId: Int!, movieId: Int!): FavoriteMovie!
+  addUserMovie(userId: Int!, movieId: Int!): FavoriteMovie!
+  removeUserMovie(userId: Int!, movieId: Int!): String!
+  toggleUserMovie(userId: Int!, movieId: Int!): FavoriteMovie!
+  addUserGenre(userId: Int!, genreId: Int!): FavoriteGenre!
+  removeUserGenre(userId: Int!, genreId: Int!): String!
 }
