+++ src/user/user.resolver.ts
@@ -7,34 +7,11 @@ import { GraphQLString } from 'graphql';
 export class UserResolver {
   constructor(private userService: UserService) {}
 
-  @Query(() => [FavoriteGenre])
-  async getUserGenres(
-    @Args('userId', { type: () => Int }) userId: number,
-  ) {
-    return (await this.userService.findUserGenres(userId)) || [];
-  }
-
   @Query(() => [FavoriteMovie])
   async getUserMovies(
     @Args('userId', { type: () => Int }) userId: number,
   ) {
     return await this.userService.findUserMovies(userId);
-  }
-
-  @Mutation(() => FavoriteGenre)
-  async addUserGenre(
-    @Args('userId', { type: () => Int }) userId: number,
-    @Args('genreId', { type: () => Int }) genreId: number,
-  ) {
-    return await this.userService.addUserGenre(userId, genreId);
-  }
-
-  @Mutation(() => GraphQLString)
-  async removeUserGenre(
-    @Args('userId', { type: () => Int }) userId: number,
-    @Args('genreId', { type: () => Int }) genreId: number,
-  ) {
-    return await this.userService.removeUserGenre(userId, genreId);
   }
 
   @Mutation(() => FavoriteMovie)
